Add max_concurrent and max_per_second config options
This adds the max_concurrent and max_per_second config options, both
which affect client side ratelimiting of AWS API requests.

These settings are useful in cases where you need to ensure the server
side ratelimiting won't be triggered. Since AWS shares the API quotas
across all services in an account a query against the environment from
one user can potentially cause RequestLimitExceeded error responses to
be returned from the API to an unrelated service.

In general applications should handle this error code gracefully and
in an ideal world this would not cause any problems. However it's
not uncommon to come accross production apps that will crash when this
happens. While you likely want to fix this issue in your code, it
is still useful to be able to limit requests when you want to be
extra careful in a production environment that you are not familar
with.

The max_concurrent setting here will prevent at most X number of
requests from being in flight at the same time. Limiting concurrent
requests should prevent any bursty request patterns, going through
the alloted requests per second all at the same time.

Setting max_per_second will allow X number of requests to be made
every second. Since AWS does not offer any metrics on the current
API quota, it is difficult to know what this needs to be set to,
you can however monitor CloudTrail for failed requests and adjust
this as needed.