fix turbot/steampipe-plugin-aws#736
This is limited to allowing SSO to auto run. Work needs to be done to have an aggregator only run this on the first child. As you only need aws sso login run one time.  Currently, as it initiates all the plugins as once you end up with N tabs in your browser opening.

If however you have run aws sso login, or  you just close and re-open steampipe it will not ask again (unless the cred is expired).